Update validation and create new api routes
