+++ pages/api/forms/[id]/delete.ts
@@ -1,5 +1,5 @@
-import { withApiAuthRequired } from '@auth0/nextjs-auth0'
 import { NextApiRequest, NextApiResponse } from 'next'
+import { getSession, withApiAuthRequired } from '@auth0/nextjs-auth0'
 
 const INSTANCE = process.env.HARPERDB_URL
 const TOKEN = process.env.HARPERDB_TOKEN
@@ -8,8 +8,15 @@ export default withApiAuthRequired(async function handler(
   req: NextApiRequest,
   res: NextApiResponse
 ) {
-  if (req.method !== 'GET') {
+  if (req.method !== 'DELETE') {
+    res.setHeader('Allow', ['DELETE'])
     res.status(405).end(`Method ${req.method} not allowed`)
+  }
+  const { id } = req.query
+  const { user }: any = getSession(req, res)
+
+  if (user?.sub !== req.body.workspace) {
+    res.status(403).end(`You're not permitted to complete this action`)
   }
 
   const request = await fetch(`${INSTANCE}`, {
@@ -21,7 +28,7 @@ export default withApiAuthRequired(async function handler(
     },
     body: JSON.stringify({
       operation: 'sql',
-      sql: `SELECT * FROM typiform.forms WHERE workspace='${req.query.id}' ORDER BY __createdtime__ ASC`,
+      sql: `DELETE FROM typiform.forms WHERE id='${id}' AND workspace='${user?.sub}'`,
     }),
   })
   const response = await request.json()
