+++ pages/api/forms/[id]/update.ts
@@ -1,5 +1,5 @@
-import { withApiAuthRequired } from '@auth0/nextjs-auth0'
 import { NextApiRequest, NextApiResponse } from 'next'
+import { getSession, withApiAuthRequired } from '@auth0/nextjs-auth0'
 
 const INSTANCE = process.env.HARPERDB_URL
 const TOKEN = process.env.HARPERDB_TOKEN
@@ -8,10 +8,18 @@ export default withApiAuthRequired(async function handler(
   req: NextApiRequest,
   res: NextApiResponse
 ) {
-  if (req.method !== 'GET') {
+  if (req.method !== 'PATCH') {
+    res.setHeader('Allow', ['PATCH'])
     res.status(405).end(`Method ${req.method} not allowed`)
   }
 
+  const { user }: any = getSession(req, res)
+
+  if (user?.sub !== req.body.workspace) {
+    res.status(403).end(`You're not permitted to complete this action`)
+  }
+
+  const body = JSON.parse(req.body)
   const request = await fetch(`${INSTANCE}`, {
     method: 'POST',
     redirect: 'follow',
@@ -20,10 +28,12 @@ export default withApiAuthRequired(async function handler(
       Authorization: `Basic ${TOKEN}`,
     },
     body: JSON.stringify({
-      operation: 'sql',
-      sql: `SELECT * FROM typiform.forms WHERE workspace='${req.query.id}' ORDER BY __createdtime__ ASC`,
+      operation: 'update',
+      schema: 'typiform',
+      table: 'forms',
+      records: [body],
     }),
   })
   const response = await request.json()
-  res.status(200).json(response)
+  res.status(200).json({ id: response.update_hashes[0] })
 })
